refactor(ide): extract IdeBreadcrumb and IdeReviewFocusStrip from ide-shell.ts
Extract two components from the 924-line ide-shell.ts to reduce file size
and separate concerns:

- IdeBreadcrumb (+ FILE_ICONS) → ide-breadcrumb.ts (~130 lines)
- IdeReviewFocusStrip → ide-review-focus-strip.ts (~20 lines)

Move shared constants IDE_LAYER_LABELS and REVIEW_FOCUS_LAYERS to
ide-toolbar.ts (the SSOT for IDE layer definitions) to avoid circular
dependencies between the extracted modules.

ide-shell.ts shrinks from 924 to 775 lines.
